osx: move CPU code in new cpu.c module
